Move to existing file
