+++ public/app/features/provisioning/guards.ts
@@ -1,3 +1,5 @@
+import { RepoType, RepoTypeDisplay } from './Wizard/types';
+
 export interface HttpError extends Error {
   status?: number;
 }
@@ -8,4 +10,11 @@ export function isSupportedGitProvider(provider: string): provider is 'github' |
 
 export function isHttpError(err: unknown): err is HttpError {
   return err instanceof Error && 'status' in err;
+}
+
+export function isValidRepoType(repoType: string | undefined): repoType is RepoType {
+  if (typeof repoType !== 'string') {
+    return false;
+  }
+  return repoType in RepoTypeDisplay;
 }
