--- public/app/features/provisioning/utils/guards.ts
@@ -1,9 +1,0 @@
-import { RepoType, RepoTypeDisplay } from '../Wizard/types';
-
-export function isValidRepoType(repoType: string | undefined): repoType is RepoType {
-  if (typeof repoType !== 'string') {
-    return false;
-  }
-  return repoType in RepoTypeDisplay;
-}
-
